Adding pauses recompute wearing time in entryDetails, but code is ugly

+++ app/src/main/java/com/lubenard/oring_reminder/ui/MainFragment.java
@@ -125,7 +125,7 @@ public class MainFragment extends Fragment {
     private void createNewEntry() {
         EditEntryFragment fragment = new EditEntryFragment(getContext());
         Bundle bundle = new Bundle();
-        bundle.putInt("entryId", -1);
+        bundle.putLong("entryId", -1);
         fragment.setArguments(bundle);
         getActivity().getSupportFragmentManager().beginTransaction()
                 .replace(android.R.id.content, fragment, null)
